update README.md, move snippets to metrics part
