+++ backend/objstore/objstore_test.go
@@ -1,16 +1,24 @@
-// Copyright © 2018 by PACE Telematics GmbH. All rights reserved.
-// Created at 2018/09/12 by Vincent Landgraf
-
 package objstore
 
 import (
+	"github.com/minio/minio-go/v6"
 	"github.com/stretchr/testify/assert"
 	"github.com/stretchr/testify/require"
 	"testing"
 )
 
-func TestObjStoreClient(t *testing.T) {
-	c, err := Client()
+func TestClient(t *testing.T) {
+	client, err := Client()
+
 	require.NoError(t, err)
-	assert.NotNil(t, c)
+	assert.NotNil(t, client)
+}
+
+func TestCustomClient(t *testing.T) {
+	client, err := CustomClient("example.org", &minio.Options{
+		Region: "eu-central-1",
+	})
+
+	require.NoError(t, err)
+	assert.NotNil(t, client)
 }
